refactor - change MA to new model type not new data type
