+++ bmds/constants.py
@@ -2,11 +2,10 @@
 
 DICHOTOMOUS = "D"
 DICHOTOMOUS_CANCER = "DC"
-DICHOTOMOUS_MA = "DCMA"
 CONTINUOUS = "C"
 CONTINUOUS_INDIVIDUAL = "CI"
-DTYPES = (DICHOTOMOUS, DICHOTOMOUS_CANCER, DICHOTOMOUS_MA, CONTINUOUS, CONTINUOUS_INDIVIDUAL)
-DICHOTOMOUS_DTYPES = (DICHOTOMOUS, DICHOTOMOUS_MA, DICHOTOMOUS_CANCER)
+DTYPES = (DICHOTOMOUS, DICHOTOMOUS_CANCER, CONTINUOUS, CONTINUOUS_INDIVIDUAL)
+DICHOTOMOUS_DTYPES = (DICHOTOMOUS, DICHOTOMOUS_CANCER)
 CONTINUOUS_DTYPES = (CONTINUOUS, CONTINUOUS_INDIVIDUAL)
 
 # bmds versions
